add : middleware to user logout

+++ app/Http/Controllers/AuthController.php
@@ -56,6 +56,12 @@ class AuthController extends Controller
 
     public function logout(Request $request): JsonResponse
     {
+        if (!$request->user()) {
+            return response()->json([
+                "errors" => ["message" => ["Unauthorized"]]
+            ], 401);
+        }
+
         // $request->user()->tokens()->delete(); delete token from all device 
         $request->user()->currentAccessToken()->delete();
